Fixed new Life360 API Auth
Changed code to work with new API requirements as well as reduced number of calls to the api which led to revoked access. Additionally added 4 new fields to the logged data (inTransit, Speed, Start Timestamp, End Timestamp. An update in the future will add functionality to visualizer.py to view this data.