Email function is complete!
Mailer.py will send out an email when called that goes to all your life360 participants with a top speed record as well as the distance that everyone has traveled in the time of which your logs have been running. Soon I'll release code into the logger360.py that will auto run this code once a week at a configurable time.